requirements.txt
# Core dependencies
streamlit==1.41.0
google-cloud-bigquery==3.17.1
python-dotenv==1.0.0

# Map and visualization
folium==0.15.1
streamlit-folium==0.15.0

# UI enhancements
extra-streamlit-components==0.1.60

# Data processing
pandas==2.1.3
numpy==1.24.3

# For the Q&A service (if you're implementing one)
# transformers==4.36.0  # later for LLM
# requests==2.31.0      # For API calls